refactor: move getSessionStore to shared folder
