+++ packages/bottender/src/index.ts
@@ -12,7 +12,7 @@ export { bottender };
 export { default as Bot } from './bot/Bot';
 export { Connector } from './bot/Connector';
 export { default as Context } from './context/Context';
-export { default as getSessionStore } from './getSessionStore';
+export { default as getSessionStore } from './shared/getSessionStore';
 export { default as getClient } from './shared/getClient';
 
 /* Action */
