+++ packages/bottender/src/shared/getSessionStore.ts
@@ -1,7 +1,8 @@
 import warning from 'warning';
 
-import SessionStore from './session/SessionStore';
-import getBottenderConfig from './shared/getBottenderConfig';
+import SessionStore from '../session/SessionStore';
+
+import getBottenderConfig from './getBottenderConfig';
 
 function getSessionStore(): SessionStore {
   const { session } = getBottenderConfig();
@@ -12,7 +13,7 @@ function getSessionStore(): SessionStore {
 
   switch (sessionDriver) {
     case 'memory': {
-      const MemorySessionStore = require('./session/MemorySessionStore')
+      const MemorySessionStore = require('../session/MemorySessionStore')
         .default;
 
       return new MemorySessionStore(
@@ -21,7 +22,7 @@ function getSessionStore(): SessionStore {
       );
     }
     case 'file': {
-      const FileSessionStore = require('./session/FileSessionStore').default;
+      const FileSessionStore = require('../session/FileSessionStore').default;
 
       return new FileSessionStore(
         storesConfig.file || {},
@@ -29,7 +30,7 @@ function getSessionStore(): SessionStore {
       );
     }
     case 'mongo': {
-      const MongoSessionStore = require('./session/MongoSessionStore').default;
+      const MongoSessionStore = require('../session/MongoSessionStore').default;
 
       return new MongoSessionStore(
         storesConfig.mongo || {},
@@ -37,7 +38,7 @@ function getSessionStore(): SessionStore {
       );
     }
     case 'redis': {
-      const RedisSessionStore = require('./session/RedisSessionStore').default;
+      const RedisSessionStore = require('../session/RedisSessionStore').default;
 
       return new RedisSessionStore(
         storesConfig.redis || {},
@@ -57,7 +58,7 @@ function getSessionStore(): SessionStore {
         false,
         `Received unknown driver: ${sessionDriver}, so fallback it to \`memory\` driver.`
       );
-      const MemorySessionStore = require('./session/MemorySessionStore')
+      const MemorySessionStore = require('../session/MemorySessionStore')
         .default;
 
       return new MemorySessionStore(
